feat: 优化前端结构
