+++ frontend/index.html
@@ -13,7 +13,7 @@
     <link rel="manifest" href="/manifest.json" />
     <title>量化选股系统</title>
   </head>
-  <body>
+  <body data-theme="dark">
     <div id="root"></div>
     <script type="module" src="/src/index.tsx"></script>
   </body>
