+++ frontend/src/components/common/ThemeToggle.tsx
@@ -1,0 +1,35 @@
+import React, { useState, useEffect } from 'react';
+import { SunIcon, MoonIcon } from '@heroicons/react/24/outline';
+
+const ThemeToggle: React.FC = () => {
+  const [isDark, setIsDark] = useState(true);
+
+  useEffect(() => {
+    // 检查当前主题
+    const currentTheme = document.documentElement.getAttribute('data-theme') || 'dark';
+    setIsDark(currentTheme === 'dark');
+  }, []);
+
+  const toggleTheme = () => {
+    const newTheme = isDark ? 'light' : 'dark';
+    document.documentElement.setAttribute('data-theme', newTheme);
+    document.body.setAttribute('data-theme', newTheme);
+    setIsDark(!isDark);
+  };
+
+  return (
+    <button
+      onClick={toggleTheme}
+      className="btn btn-ghost btn-sm"
+      title={isDark ? '切换到浅色主题' : '切换到深色主题'}
+    >
+      {isDark ? (
+        <SunIcon className="w-4 h-4" />
+      ) : (
+        <MoonIcon className="w-4 h-4" />
+      )}
+    </button>
+  );
+};
+
+export default ThemeToggle;
