+++ frontend/src/pages/QuantitativeSelection.tsx
@@ -1,14 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import {
-  PlusIcon,
   PlayIcon,
-  AdjustmentsHorizontalIcon,
   ChartBarIcon,
-  ListBulletIcon,
   Cog6ToothIcon,
   StarIcon,
-  ArrowPathIcon,
-  ExclamationTriangleIcon,
+  AdjustmentsHorizontalIcon,
   BeakerIcon
 } from '@heroicons/react/24/outline';
 
@@ -17,10 +13,14 @@ import FactorLibrary from '../components/FactorLibrary';
 import StrategyConfigManager from '../components/StrategyConfigManager';
 import WeightConfigPanel from '../components/WeightConfigPanel';
 import FactorAnalysisPanel from '../components/FactorAnalysisPanel';
+import SelectionWizard from '../components/quantitative/SelectionWizard';
+import SelectedFactorsOverview from '../components/quantitative/SelectedFactorsOverview';
+import StrategyResultModal from '../components/quantitative/StrategyResultModal';
+import MessageAlert from '../components/dashboard/MessageAlert';
 
 // API导入
 import {
-  factorApi, // 替换 builtinFactorApi
+  factorApi,
   strategyConfigApi,
   strategyApi,
   weightApi,
@@ -220,18 +220,7 @@ const QuantitativeSelection: React.FC = () => {
   return (
     <div className="space-y-6">
       {/* 消息提示 */}
-      {message && (
-        <div className={`alert ${
-          message.type === 'success' ? 'alert-success' :
-          message.type === 'error' ? 'alert-error' :
-          message.type === 'warning' ? 'alert-warning' :
-          'alert-info'
-        } shadow-lg`}>
-          <div>
-            <span>{message.text}</span>
-          </div>
-        </div>
-      )}
+      <MessageAlert message={message} />
 
       {/* 顶部导航 */}
       <div className="flex flex-col lg:flex-row justify-between items-start lg:items-center gap-4">
@@ -299,88 +288,16 @@ const QuantitativeSelection: React.FC = () => {
       </div>
 
       {/* 已选因子概览 */}
-      {selectedFactors.length > 0 && (
-        <div className="card bg-base-100 shadow-lg">
-          <div className="card-body p-4">
-            <div className="flex justify-between items-center mb-3">
-              <h3 className="font-semibold">
-                已选择因子 ({selectedFactors.length})
-                {currentStrategy && (
-                  <span className="text-sm text-base-content/60 ml-2">
-                    来自策略: {currentStrategy.name}
-                  </span>
-                )}
-              </h3>
-              <div className="flex items-center gap-2">
-                <span className={`text-sm ${isWeightValid() ? 'text-success' : 'text-warning'}`}>
-                  总权重: {(getTotalWeight() * 100).toFixed(1)}%
-                </span>
-                {!isWeightValid() && (
-                  <ExclamationTriangleIcon className="w-4 h-4 text-warning" />
-                )}
-              </div>
-            </div>
-
-            <div className="grid grid-cols-1 lg:grid-cols-2 xl:grid-cols-3 gap-3">
-              {selectedFactors.map((factor, index) => (
-                <div
-                  key={`${factor.factor_id}-${index}`}
-                  className={`p-3 border rounded-lg ${
-                    factor.is_enabled ? 'border-primary bg-primary/5' : 'border-base-300 bg-base-200'
-                  }`}
-                >
-                  <div className="flex justify-between items-start mb-2">
-                    <h4 className="font-medium text-sm truncate">{factor.name}</h4>
-                    <div className="flex gap-1">
-                      <input
-                        type="checkbox"
-                        className="checkbox checkbox-sm"
-                        checked={factor.is_enabled}
-                        onChange={() => handleFactorToggle(factor.factor_id)}
-                      />
-                      <button
-                        className="btn btn-ghost btn-xs"
-                        onClick={() => handleRemoveFactor(factor.factor_id)}
-                      >
-                        ×
-                      </button>
-                    </div>
-                  </div>
-
-                  <div className="flex items-center gap-2">
-                    <span className="text-xs">权重:</span>
-                    <input
-                      type="range"
-                      min="0"
-                      max="1"
-                      step="0.01"
-                      value={factor.weight}
-                      onChange={(e) => handleFactorWeightChange(factor.factor_id, parseFloat(e.target.value))}
-                      className="range range-xs flex-1"
-                      disabled={!factor.is_enabled}
-                    />
-                    <span className="text-xs w-12 text-right">
-                      {((factor.weight ?? 0) * 100).toFixed(0)}%
-                    </span>
-                  </div>
-                </div>
-              ))}
-            </div>
-
-            {!isWeightValid() && (
-              <div className="mt-3 flex gap-2">
-                <button
-                  className="btn btn-warning btn-sm"
-                  onClick={handleNormalizeWeights}
-                >
-                  <ArrowPathIcon className="w-4 h-4" />
-                  标准化权重
-                </button>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
+      <SelectedFactorsOverview
+        selectedFactors={selectedFactors}
+        currentStrategy={currentStrategy}
+        isWeightValid={isWeightValid}
+        getTotalWeight={getTotalWeight}
+        onRemoveFactor={handleRemoveFactor}
+        onFactorWeightChange={handleFactorWeightChange}
+        onFactorToggle={handleFactorToggle}
+        onNormalizeWeights={handleNormalizeWeights}
+      />
 
       {/* 主要内容区域 */}
       <div className="min-h-[600px]">
@@ -400,192 +317,17 @@ const QuantitativeSelection: React.FC = () => {
         )}
 
         {activeTab === 'selection-wizard' && (
-          <div className="card bg-base-100 shadow-lg">
-            <div className="card-body">
-              <h2 className="card-title mb-4">选股向导</h2>
-              
-              {selectedFactors.length === 0 ? (
-                <div className="text-center py-12">
-                  <div className="text-base-content/60 mb-4">
-                    请先从因子库中选择因子，或加载一个策略配置
-                  </div>
-                  <div className="flex gap-2 justify-center">
-                    <button
-                      className="btn btn-primary"
-                      onClick={() => setActiveTab('factor-library')}
-                    >
-                      <ChartBarIcon className="w-5 h-5" />
-                      浏览因子库
-                    </button>
-                    <button
-                      className="btn btn-outline"
-                      onClick={() => setActiveTab('strategy-config')}
-                    >
-                      <Cog6ToothIcon className="w-5 h-5" />
-                      加载策略
-                    </button>
-                  </div>
-                </div>
-              ) : (
-                <div className="space-y-6">
-                  {/* 执行配置 */}
-                  <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
-                    <div className="form-control">
-                      <label className="label">
-                        <span className="label-text">最大结果数</span>
-                      </label>
-                      <input
-                        type="number"
-                        className="input input-bordered"
-                        value={executeForm.max_results}
-                        onChange={(e) => setExecuteForm({
-                          ...executeForm,
-                          max_results: parseInt(e.target.value) || 50
-                        })}
-                        min="1"
-                        max="500"
-                      />
-                    </div>
-
-                    <div className="form-control">
-                      <label className="label">
-                        <span className="label-text">最小市值 (万元)</span>
-                      </label>
-                      <input
-                        type="number"
-                        className="input input-bordered"
-                        value={executeForm.filters.min_market_cap / 10000}
-                        onChange={(e) => setExecuteForm({
-                          ...executeForm,
-                          filters: {
-                            ...executeForm.filters,
-                            min_market_cap: (parseInt(e.target.value) || 100) * 10000
-                          }
-                        })}
-                        placeholder="100"
-                      />
-                    </div>
-                  </div>
-
-                  {/* 过滤条件 */}
-                  <div>
-                    <label className="label">
-                      <span className="label-text">过滤条件</span>
-                    </label>
-                    <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
-                      <div className="form-control">
-                        <label className="label cursor-pointer">
-                          <span className="label-text">排除ST股</span>
-                          <input
-                            type="checkbox"
-                            className="checkbox"
-                            checked={executeForm.filters.exclude_st}
-                            onChange={(e) => setExecuteForm({
-                              ...executeForm,
-                              filters: {
-                                ...executeForm.filters,
-                                exclude_st: e.target.checked
-                              }
-                            })}
-                          />
-                        </label>
-                      </div>
-
-                      <div className="form-control">
-                        <label className="label cursor-pointer">
-                          <span className="label-text">排除新股</span>
-                          <input
-                            type="checkbox"
-                            className="checkbox"
-                            checked={executeForm.filters.exclude_new_stock}
-                            onChange={(e) => setExecuteForm({
-                              ...executeForm,
-                              filters: {
-                                ...executeForm.filters,
-                                exclude_new_stock: e.target.checked
-                              }
-                            })}
-                          />
-                        </label>
-                      </div>
-
-                      <div className="form-control">
-                        <label className="label">
-                          <span className="label-text">最大市盈率</span>
-                        </label>
-                        <input
-                          type="number"
-                          className="input input-bordered"
-                          value={executeForm.filters.max_pe}
-                          onChange={(e) => setExecuteForm({
-                            ...executeForm,
-                            filters: {
-                              ...executeForm.filters,
-                              max_pe: parseInt(e.target.value) || 100
-                            }
-                          })}
-                          placeholder="100"
-                        />
-                      </div>
-                    </div>
-                  </div>
-
-                  {/* 权重验证和建议 */}
-                  {!isWeightValid() && (
-                    <div className="alert alert-warning">
-                      <ExclamationTriangleIcon className="w-5 h-5" />
-                      <div>
-                        <div className="font-semibold">权重配置需要调整</div>
-                        <div className="text-sm">
-                          当前总权重为 {(getTotalWeight() * 100).toFixed(1)}%，建议调整为100%以获得最佳结果。
-                        </div>
-                      </div>
-                      <div>
-                        <button
-                          className="btn btn-warning btn-sm"
-                          onClick={handleNormalizeWeights}
-                        >
-                          自动调整
-                        </button>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* 执行按钮 */}
-                  <div className="flex gap-4 justify-center">
-                    <button
-                      className="btn btn-outline"
-                      onClick={() => setActiveTab('weight-config')}
-                    >
-                      <AdjustmentsHorizontalIcon className="w-5 h-5" />
-                      高级权重配置
-                    </button>
-                    <button
-                      className="btn btn-outline"
-                      onClick={() => setActiveTab('factor-analysis')}
-                    >
-                      <BeakerIcon className="w-5 h-5" />
-                      因子分析
-                    </button>
-                    <button
-                      className={`btn btn-primary btn-lg ${executing ? 'loading' : ''}`}
-                      onClick={() => handleExecuteStrategy()}
-                      disabled={executing || selectedFactors.length === 0}
-                    >
-                      {executing ? (
-                        '执行中...'
-                      ) : (
-                        <>
-                          <PlayIcon className="w-5 h-5" />
-                          开始选股
-                        </>
-                      )}
-                    </button>
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
+          <SelectionWizard
+            selectedFactors={selectedFactors}
+            executeForm={executeForm}
+            executing={executing}
+            isWeightValid={isWeightValid}
+            getTotalWeight={getTotalWeight}
+            onExecuteStrategy={() => handleExecuteStrategy()}
+            onNormalizeWeights={handleNormalizeWeights}
+            onSetExecuteForm={setExecuteForm}
+            onSetActiveTab={setActiveTab}
+          />
         )}
 
         {activeTab === 'weight-config' && (
@@ -603,95 +345,12 @@ const QuantitativeSelection: React.FC = () => {
         )}
       </div>
 
-
-
       {/* 策略执行结果模态框 */}
-      {showResultModal && (
-        <div className="modal modal-open">
-          <div className="modal-box w-11/12 max-w-6xl">
-            <h3 className="font-bold text-lg mb-4">
-              选股结果 ({strategyResults.length} 只股票)
-            </h3>
-
-            {strategyResults.length > 0 ? (
-              <div className="overflow-x-auto">
-                <table className="table table-sm">
-                  <thead>
-                    <tr>
-                      <th>排名</th>
-                      <th>股票代码</th>
-                      <th>股票名称</th>
-                      <th>综合得分</th>
-                      <th>当前价格</th>
-                      <th>涨跌幅</th>
-                      <th>行业</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {strategyResults.map((result, index) => (
-                      <tr key={result.symbol} className={index < 10 ? 'bg-success/10' : ''}>
-                        <td className="font-medium">#{index + 1}</td>
-                        <td className="font-mono">{result.symbol}</td>
-                        <td>{result.name}</td>
-                        <td>
-                          <div className="badge badge-primary">
-                            {result.score.toFixed(2)}
-                          </div>
-                        </td>
-                        <td>¥{result.price.toFixed(2)}</td>
-                        <td>
-                          <span className={`font-medium ${
-                            result.changePercent >= 0 ? 'text-success' : 'text-error'
-                          }`}>
-                            {result.changePercent >= 0 ? '+' : ''}{result.changePercent.toFixed(2)}%
-                          </span>
-                        </td>
-                        <td>{result.industry || '-'}</td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            ) : (
-              <div className="text-center py-12">
-                <div className="text-base-content/60">
-                  没有找到符合条件的股票
-                </div>
-              </div>
-            )}
-
-            <div className="modal-action">
-              <button
-                className="btn"
-                onClick={() => setShowResultModal(false)}
-              >
-                关闭
-              </button>
-              {strategyResults.length > 0 && (
-                <button
-                  className="btn btn-primary"
-                  onClick={() => {
-                    // 导出结果逻辑
-                    const csv = strategyResults.map((result, index) => 
-                      `${index + 1},${result.symbol},${result.name},${result.score.toFixed(2)},${result.price.toFixed(2)},${result.changePercent.toFixed(2)}%`
-                    ).join('\n');
-                    
-                    const blob = new Blob([`排名,代码,名称,得分,价格,涨跌幅\n${csv}`], { type: 'text/csv' });
-                    const url = URL.createObjectURL(blob);
-                    const a = document.createElement('a');
-                    a.href = url;
-                    a.download = `selection_result_${new Date().toISOString().split('T')[0]}.csv`;
-                    a.click();
-                    URL.revokeObjectURL(url);
-                  }}
-                >
-                  导出结果
-                </button>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
+      <StrategyResultModal
+        visible={showResultModal}
+        results={strategyResults}
+        onClose={() => setShowResultModal(false)}
+      />
     </div>
   );
 };
